Format SHOW queries and send queries as body
For very long queries like `CREATE` with large amounts
of columns, we run into limitations of the parameter length.
As it is only necessary to put the query into the URL for
INSERT statements, the default is now to send them in the body.

Further SHOW statements use the FORMAT statement for the
configured codec.